Rename to netty3
