+++ agent/agent-plugins/httpclient-netty3/pom.xml
@@ -7,7 +7,7 @@
     <version>1.0-SNAPSHOT</version>
     <relativePath>../../pom.xml</relativePath>
   </parent>
-  <artifactId>agent-plugin-httpclient-netty</artifactId>
+  <artifactId>agent-plugin-httpclient-netty3</artifactId>
 
   <dependencies>
     <dependency>
@@ -31,7 +31,7 @@
         <configuration>
           <archive>
             <manifestEntries>
-              <Plugin-Class>org.bithon.agent.plugin.httpclient.netty.NettyHttpClientPlugin
+              <Plugin-Class>org.bithon.agent.plugin.httpclient.netty3.NettyHttpClientPlugin
               </Plugin-Class>
             </manifestEntries>
           </archive>
